Copied from Perforce
 Change: 185118
 ServerID: perforce.ravenbrook.com

+++ mps/tool/testrun.bat
@@ -61,9 +61,9 @@ set TEST_COUNT=0
 set PASS_COUNT=0
 set FAIL_COUNT=0
 set SEPARATOR=----------------------------------------
-set LOGDIR=%TMP%\mps-%VARIETY%-log
+set LOGDIR=%TMP%\mps-%PFM%-%VARIETY%-log
 echo Logging test output to %LOGDIR%
-rmdir /q /s %LOGDIR%
+if exist %LOGDIR% rmdir /q /s %LOGDIR%
 mkdir %LOGDIR%
 
 if "%1"=="" call :run_tests %ALL_TEST_CASES%
